+++ attributes/host.rb
@@ -2,6 +2,6 @@
 # Attributes:: host
 
 
-default['op5_manage']['hosts'] = []
+default['op5_manage']['hosts'] = {}
 
 # Attributes for Test Kitchen are defined in .kitchen.yml