+++ attributes/host_downtime.rb
@@ -2,6 +2,6 @@
 # Attributes:: host_downtime
 
 
-default['op5_manage']['host_downtimes'] = []
+default['op5_manage']['host_downtimes'] = {}
 
 # Attributes for Test Kitchen are defined in .kitchen.yml