+++ attributes/node.rb
@@ -3,15 +3,18 @@
 
 
 default['op5_manage']['node'] = {
-    action:           :create
+    'centos'    => { 'template' => 'default-host-template', 'hostgroups' => { 'Linux servers' => true }},
+    'redhat'    => { 'template' => 'default-host-template', 'hostgroups' => { 'Linux servers' => true }},
+    'suse'      => { 'template' => 'default-host-template', 'hostgroups' => { 'Linux servers' => true }},
+    action:     :create
 }
 
 
 # If true, initial downtime for host will be set on server provisioning
-default['op5_manage']['initial_downtime']['enabled'] = true
+default['op5_manage']['node']['initial_downtime']['enabled'] = true
 
 # If false, server is newly provisioned
-default['op5_manage']['initial_downtime']['scheduled'] = false
+default['op5_manage']['node']['initial_downtime']['scheduled'] = false
 
 # Duration of initial downtime in hours
-default['op5_manage']['initial_downtime']['duration'] = 336     # Two weeks
+default['op5_manage']['node']['initial_downtime']['duration'] = 24