+++ attributes/services.rb
@@ -2,6 +2,6 @@
 # Attributes:: service
 
 
-default['op5_manage']['services'] = []
+default['op5_manage']['services'] = {}
 
 # Attributes for Test Kitchen are defined in .kitchen.yml