+++ metadata.rb
@@ -7,13 +7,14 @@ issues_url        'https://github.com/SlashDashAndCash/op5_manage/issues'
 description       'Manage op5 hosts and services'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 
-version '0.8.5'
+version '1.0.0'
 
 # Weired release numbers for supermarket
 supports 'centos', '>= 6.0.0'
 supports 'redhat', '>= 6.0.0'
 supports 'suse',   '>= 11.0.0'
 
-chef_version       '>= 12.6.0'
+chef_version       '>= 12.14.0'
 
 depends 'chef-vault'
+