+++ providers/host.rb
@@ -37,6 +37,7 @@ action :create do
     result = op5manage.create_host(host_config)
     if result
       Chef::Log.info "Sucessfully created host #{new_resource.name}"
+      node.run_state['config_changed'] = true
       new_resource.updated_by_last_action(true)
     else
       Chef::Log.error "Couldn't create host #{new_resource.name}"
@@ -50,6 +51,7 @@ action :create do
       result = op5manage.update_host(host_config)
       if result
         Chef::Log.info "Sucessfully updated host #{new_resource.name}"
+        node.run_state['config_changed'] = true
         new_resource.updated_by_last_action(true)
       else
         Chef::Log.error "Couldn't update host #{new_resource.name}"
@@ -76,6 +78,7 @@ action :remove do
     result = op5manage.remove_host(new_resource.name)
     if result
       Chef::Log.info "Sucessfully removed host #{new_resource}"
+      node.run_state['config_changed'] = true
       new_resource.updated_by_last_action(true)
     else
       Chef::Log.error "Couldn't remove host #{new_resource}"
@@ -99,9 +102,9 @@ def create_host_config
   host_config['check_freshness']                = new_resource.check_freshness              unless new_resource.check_freshness.nil?
   host_config['check_interval']                 = new_resource.check_interval               unless new_resource.check_interval.nil?
   host_config['check_period']                   = new_resource.check_period                 unless new_resource.check_period.nil?
-  host_config['children']                       = new_resource.children                     unless new_resource.children.nil?
-  host_config['contact_groups']                 = new_resource.contact_groups               unless new_resource.contact_groups.nil?
-  host_config['contacts']                       = new_resource.contacts                     unless new_resource.contacts.nil?
+  host_config['children']                       = new_resource.children.select{|key, flag| flag == true }.keys                unless new_resource.children.nil?
+  host_config['contact_groups']                 = new_resource.contact_groups.select{|key, flag| flag == true }.keys          unless new_resource.contact_groups.nil?
+  host_config['contacts']                       = new_resource.contacts.select{|key, flag| flag == true }.keys                unless new_resource.contacts.nil?
   host_config['custom_variable']                = new_resource.custom_variable              unless new_resource.custom_variable.nil?
   host_config['display_name']                   = new_resource.display_name                 unless new_resource.display_name.nil?
   host_config['event_handler']                  = new_resource.event_handler                unless new_resource.event_handler.nil?
@@ -109,10 +112,10 @@ def create_host_config
   host_config['event_handler_enabled']          = new_resource.event_handler_enabled        unless new_resource.event_handler_enabled.nil?
   host_config['first_notification_delay']       = new_resource.first_notification_delay     unless new_resource.first_notification_delay.nil?
   host_config['flap_detection_enabled']         = new_resource.flap_detection_enabled       unless new_resource.flap_detection_enabled.nil?
-  host_config['flap_detection_options']         = new_resource.flap_detection_options       unless new_resource.flap_detection_options.nil?
+  host_config['flap_detection_options']         = new_resource.flap_detection_options.select{|key, flag| flag == true }.keys  unless new_resource.flap_detection_options.nil?
   host_config['freshness_threshold']            = new_resource.freshness_threshold          unless new_resource.freshness_threshold.nil?
   host_config['high_flap_threshold']            = new_resource.high_flap_threshold          unless new_resource.high_flap_threshold.nil?
-  host_config['hostgroups']                     = new_resource.hostgroups                   unless new_resource.hostgroups.nil?
+  host_config['hostgroups']                     = new_resource.hostgroups.select{|key, flag| flag == true }.keys              unless new_resource.hostgroups.nil?
   host_config['icon_image']                     = new_resource.icon_image                   unless new_resource.icon_image.nil?
   host_config['icon_image_alt']                 = new_resource.icon_image_alt               unless new_resource.icon_image_alt.nil?
   host_config['low_flap_threshold']             = new_resource.low_flap_threshold           unless new_resource.low_flap_threshold.nil?
@@ -121,18 +124,18 @@ def create_host_config
   host_config['notes']                          = new_resource.notes                        unless new_resource.notes.nil?
   host_config['notes_url']                      = new_resource.notes_url                    unless new_resource.notes_url.nil?
   host_config['notification_interval']          = new_resource.notification_interval        unless new_resource.notification_interval.nil?
-  host_config['notification_options']           = new_resource.notification_options         unless new_resource.notification_options.nil?
+  host_config['notification_options']           = new_resource.notification_options.select{|key, flag| flag == true }.keys    unless new_resource.notification_options.nil?
   host_config['notification_period']            = new_resource.notification_period          unless new_resource.notification_period.nil?
   host_config['notifications_enabled']          = new_resource.notifications_enabled        unless new_resource.notifications_enabled.nil?
   host_config['obsess']                         = new_resource.obsess                       unless new_resource.obsess.nil?
-  host_config['parents']                        = new_resource.parents                      unless new_resource.parents.nil?
+  host_config['parents']                        = new_resource.parents.select{|key, flag| flag == true }.keys                 unless new_resource.parents.nil?
   host_config['passive_checks_enabled']         = new_resource.passive_checks_enabled       unless new_resource.passive_checks_enabled.nil?
   host_config['process_perf_data']              = new_resource.process_perf_data            unless new_resource.process_perf_data.nil?
   host_config['register']                       = new_resource.register                     unless new_resource.register.nil?
   host_config['retain_nonstatus_information']   = new_resource.retain_nonstatus_information unless new_resource.retain_nonstatus_information.nil?
   host_config['retain_status_information']      = new_resource.retain_status_information    unless new_resource.retain_status_information.nil?
   host_config['retry_interval']                 = new_resource.retry_interval               unless new_resource.retry_interval.nil?
-  host_config['stalking_options']               = new_resource.stalking_options             unless new_resource.stalking_options.nil?
+  host_config['stalking_options']               = new_resource.stalking_options.select{|key, flag| flag == true }.keys        unless new_resource.stalking_options.nil?
   host_config['statusmap_image']                = new_resource.statusmap_image              unless new_resource.statusmap_image.nil?
   host_config['template']                       = new_resource.template
   host_config['2d_coords']                      = new_resource.two_d_coords                 unless new_resource.two_d_coords.nil?