+++ resources/service.rb
@@ -29,8 +29,8 @@ attribute :check_command_args,            :kind_of => String, :default => nil
 attribute :check_freshness,               :kind_of => [ TrueClass, FalseClass ], :default => nil
 attribute :check_interval,                :kind_of => Integer, :default => nil
 attribute :check_period,                  :kind_of => String, :default => nil
-attribute :contact_groups,                :kind_of => Array,  :default => nil
-attribute :contacts,                      :kind_of => Array,  :default => nil
+attribute :contact_groups,                :kind_of => Hash,  :default => nil
+attribute :contacts,                      :kind_of => Hash,  :default => nil
 attribute :display_name,                  :kind_of => String, :default => nil
 attribute :event_handler,                 :kind_of => String, :default => nil
 attribute :event_handler_args,            :kind_of => String, :default => nil
@@ -38,7 +38,7 @@ attribute :event_handler_enabled,         :kind_of => [ TrueClass, FalseClass ],
 attribute :file_id,                       :kind_of => String, :default => nil
 attribute :first_notification_delay,      :kind_of => Integer, :default => nil
 attribute :flap_detection_enabled,        :kind_of => [ TrueClass, FalseClass ], :default => nil
-attribute :flap_detection_options,        :kind_of => Array,  :default => nil
+attribute :flap_detection_options,        :kind_of => Hash,  :default => nil
 attribute :freshness_threshold,           :kind_of => Integer, :default => nil
 attribute :high_flap_threshold,           :kind_of => Integer, :default => nil
 #Hostgroups are unsupported yet
@@ -50,7 +50,7 @@ attribute :max_check_attempts,            :kind_of => Integer, :default => nil
 attribute :notes,                         :kind_of => String, :default => nil
 attribute :notes_url,                     :kind_of => String, :default => nil
 attribute :notification_interval,         :kind_of => Integer, :default => nil
-attribute :notification_options,          :kind_of => Array,  :default => nil
+attribute :notification_options,          :kind_of => Hash,  :default => nil
 attribute :notification_period,           :kind_of => String, :default => nil
 attribute :notifications_enabled,         :kind_of => [ TrueClass, FalseClass ], :default => nil
 attribute :obsess,                        :kind_of => [ TrueClass, FalseClass ], :default => nil
@@ -60,7 +60,7 @@ attribute :process_perf_data,             :kind_of => [ TrueClass, FalseClass ],
 attribute :register,                      :kind_of => [ TrueClass, FalseClass ], :default => nil
 attribute :retain_nonstatus_information,  :kind_of => [ TrueClass, FalseClass ], :default => nil
 attribute :retain_status_information,     :kind_of => [ TrueClass, FalseClass ], :default => nil
-attribute :servicegroups,                 :kind_of => Array,  :default => nil
+attribute :servicegroups,                 :kind_of => Hash,  :default => nil
 attribute :retry_interval,                :kind_of => Integer, :default => nil
-attribute :stalking_options,              :kind_of => Array,  :default => nil
+attribute :stalking_options,              :kind_of => Hash,  :default => nil
 attribute :template,                      :kind_of => String, :default => nil